Python Web scraper/requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
aspose-cells==22.11.0
asttokens==2.2.1
attrs==22.1.0
autopep8==2.0.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.10.0
bleach==5.0.1
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==38.0.4
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
firebase-admin==6.0.1
fqdn==1.5.1
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
importlib-metadata==5.1.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
JPype1==1.4.1
jsonpointer==2.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
msgpack==1.0.4
mypy-extensions==0.4.3
mysql-connector==2.2.9
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.2
pickleshare==0.7.5
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.36
proto-plus==1.22.1
protobuf==4.21.10
psutil==5.9.4
pure-eval==0.2.2
py-bt==1.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycricbuzz==2.4
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
pywin32==305
pywinpty==2.0.9
PyYAML==5.3
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.7.1
typing_extensions==4.4.0
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.4
XlsxWriter==3.0.3
zipp==3.11.0
